requirements.txt
numpy~=2.2.6
faiss-cpu~=1.11.0
torch~=2.7.0
pillow~=11.2.1
uvicorn~=0.34.3
fastapi~=0.115.12
slowapi~=0.1.9
python-dotenv~=1.1.0
transformers~=4.52.4